Automatically generate a CEK during wrapping
In jose_jwe_wrap(), if a CEK is passed that has no "k" attribute, we
will now generate this internally. Notably, the core code will fill in
an algorithm if not specified. It is the job of wrappers to actually
generate the key.

This makes the wrapping process easier to use. It also facilitates for
ECDH-ES which requires this internal generation.

Fixes #3