Default jose ver to use stdout if not detached
When the payload is embedded in the JWS, the user probably wants to
read the payload from the JWS. So we default to output on stdout. If,
however, we are in detached mode, the user already has access to the
payload outside the JWS. So in this case, we presume the user doesn't
want any payload output.