Fix RSA SSA PSS salt length
This patch fixes the PSS signature salt length to match the
requirements of RFC 7518 (JWA), which says the following:

3.5. Digital Signature with RSASSA-PSS

This section defines the use of the RSASSA-PSS digital signature
algorithm as defined in Section 8.1 of RFC 3447 [RFC3447] with the
MGF1 mask generation function and SHA-2 hash functions, always using
the same hash function for both the RSASSA-PSS hash function and the
MGF1 hash function. The size of the salt value is the same size as
the hash function output. All other algorithm parameters use the
defaults specified in Appendix A.2.3 of RFC 3447.