Major rework to handle streaming efficiently
This commit breaks API/ABI so that we can process encryption and signing in
chunks rather than all at once. The end result should be that we can sign
or encrypt very large data without loading it all into memory at once.

The new API is tested much more extensively then in the past.

We also rework the way commands are handled in jose. Now, commands are
grouped in order to keep track of the different data types.

We have also added API and CLI documentation, including man pages.