Test the base64 decoder extensively
This test presumes that our encoder works. From there it generates
every possible input to the encoder and tracks the outputs that occur.
Then it generates every possible value to the decoder and checks each
against the valid value table to determine if the decoder should succeed
or fail. This is essentially a brute force proof that the decoder works
for every possible value.

Thanks to OpenMP we can do this in parallel without much effort.
Unfortunately, this still takes a lot of time to test (~5-6 minutes on
my system). This test is fully enabled where OpenMP is found and not
disabled (--disable-openmp). On other systems, it only runs a partial
test. Due to resource constraints, we only run the short version on
TravisCI.