configure: fix build with old compilers
Old gcc versions (gcc 4.7) do not support all warnings flags currently
hard-coded by configure.ac. In order to fix this, we import the
AX_CHECK_COMPILE_FLAG() macro from the autoconf-archive, and use it in
the configure.ac to only use warning flags when the compiler supports
them.

As a side effect, we need to add a file in m4/, so we adjust the
.gitignore file accordingly.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>