ci: add github actions
As travis as showing to be unreliable and slow, let's move the CI to
Github actions.

Note: I am not sure it supports ppc64le for now, so that wil not be
there as of now. Perhaps we can amend it in the future.

Note 2: ubuntu:bionic seems to have some issues with meson, so I left
it out also for now.

We are builing the following platforms for x86_64, with both clang and
gcc:

- fedora:rawhide
- fedora:latest
- centos:8
- debian:testing
- debian:latest
- ubuntu:devel
- ubuntu:rolling
- macos-latest