Fix JWE encryption not setting protected header correctly (#114)
cmd/jwe/enc.c - this makes the compact code the default, so the right values get added to the protected header all the time. This only fixes the jose cli.

lib/jwe.c - calling jose_jwe_hdr and passing jwe for both values will take the current protected field and the header field and combine them, then reset it as the protected field, making sure all the right values are set in the protected value. This only fixes the lib api calls.

tests/jose-jwe-enc - this updates the tests to look in the protected value instead of the header value for things.
---------

Co-authored-by: Brian France <brian@Brians-MacBook-Pro.local>