Remove the locals regexes from libjose.map.  They seem to serve no purpose (#173)
and are capturing the environ and __prognam symbols from libc on FreeBSD
causing various linking errors due to those symbols being erroneously not
exported.  Minor change in meson.build to look for build_machine vs. host_machine
because the linking is happening on the build machine and thus the linker syntax
is build machine dependent.